# Preprocessing speedup