#138 
Optimized the following functions 

| function                             | before       | after  | speedup |
|--------------------------------------|--------------|--------|---------|
| `pp.utilities.error_correct_intbc`   | 2 min        | 1 min  | 2x      |
| `pp.utilities.record_stats`          | 20 sec       | 178 ms | 140x    |
| `pp.UMI_utils.correct_umis_in_group` | 1 min 25 sec | 138 ms | 615x    |
| `pp.align_sequences` (10000 sequences) | 30 sec | 26 sec |  1.15x |
| `pp.map_utils.map_intbcs` | 6 min 54 sec | 3.47 sec | 119x |
| `pp.doublet_utils.filter_intra_doublets` | 49 sec | 361 ms | 135x |
| `pp.doublet_utils.get_intbc_set(s)` | 713 ms | 149 ms | 4x |
| `pp.doublet_utils.filter_inter_doublets` | 5 sec |  860 ms | 5x |

Other changes
* All instances of `moleculetable` have been made consistent to `molecule_table`
* Changed parameter names for `pp.utilities.error_correct_intbc` to snake-case.
* `pp.utilities.convert_bam_to_df` always returns a pandas Dataframe, and the caller must decide to write or not
* `pp.UMI_utils.correct_umis_in_group` no longer returns the total number of UMIs because that can be obtained by simply taking the number of rows in the returned dataframe.
* `pp.map_utils` is no longer a Cython extension, but a regular .py file. It was not using any Cython features.
* Changed `pp.doublet_utils.get_intbc_sets` to `pp.doublet_utils.get_intbc_set` such that the function operates on a single lineage group.
* `pp.doublet_utils` is no longer a Cython extension. It was not using any Cython features.