@Lioscro 2021-08-18
Also @mattjones315, I've noticed in `pp.map_utils.map_intbcs`, the intBC with the highest *read count* is selected. Shouldn't we be doing this by UMI?